치명타값 등 부동소수점 필요한것들 적용중 + 플레이어 리셋시 죽음처리 초기화 + takeDamage시 분기처리 버그수정

+++ Assets/02_Scripts/General/Data/Enemy/EnemyData.cs
@@ -20,9 +20,9 @@ public enum EnemyID
 public class EnemyData : BaseData
 {
     public string Name;
-    public int MaxHp;
-    public int MoveSpeed;
-    public int AttackPower;
-    public int AttackSpeed;
+    public float MaxHp;
+    public float MoveSpeed;
+    public float AttackPower;
+    public float AttackSpeed;
     public int DropGold;
 }
