+++ Assets/02_Scripts/Ingame/HP/IDamagable.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 
 public interface IDamagable
 {
-    int CurrentHp { get; }
-    int MaxHp { get; }
+    float CurrentHp { get; }
+    float MaxHp { get; }
     bool IsDead { get; }
-    void TakeDamage(int atk, bool isCritical);
+    void TakeDamage(float atk, bool isCritical);
     void Die();
 }
